Tkt 68 feature/history

+++ src/main/java/com/trillion/tikitaka/ticket/application/TicketService.java
@@ -337,9 +337,11 @@ public class TicketService {
 
         ticket.updateStatus(Ticket.Status.IN_PROGRESS);
 
-        ticket.updateManager(manager);
+        if(ticket.getManager() == null){
+            ticket.updateManager(manager);
+        }
 
-        historyService.recordHistory(ticket,manager, TicketHistory.UpdateType.TICKET_APPROVED);
+        historyService.recordHistory(ticket,manager, TicketHistory.UpdateType.STATUS_CHANGE);
 
     }
 
@@ -355,12 +357,13 @@ public class TicketService {
     public void deleteTicket(Long ticketId, CustomUserDetails userDetails) {
         Ticket ticket = ticketRepository.findById(ticketId)
                 .orElseThrow(TicketNotFoundException::new);
-
-        ticketRepository.delete(ticket);
-
         User user = userDetails.getUser();
-
-        historyService.recordHistory(ticket,user, TicketHistory.UpdateType.TICKET_DELETE);
+        User requester = userRepository.findById(ticket.getRequester().getId())
+                .orElseThrow(UserNotFoundException::new);
+        if(user.getUsername().equals(requester.getUsername()) && ticket.getStatus().equals(Ticket.Status.PENDING)) {
+            ticketRepository.delete(ticket);
+            historyService.recordHistory(ticket,user, TicketHistory.UpdateType.TICKET_DELETE);
+        }else{ throw new UnauthorizedTicketAccessException();}
 
     }
